refactor move to footballDataModel and reuse
