+++ sport/app/football/model/DotcomRenderingFootballDataModel.scala
@@ -3,10 +3,10 @@ package football.model
 import common.{CanonicalLink, Edition}
 import conf.Configuration
 import experiments.ActiveExperiments
-import football.controllers.{CompetitionFilter, FootballPage}
+import football.controllers.{CompetitionFilter, FootballPage, MatchDataAnswer, MatchPage}
 import model.dotcomrendering.DotcomRenderingUtils.{assetURL, withoutDeepNull, withoutNull}
 import model.dotcomrendering.{Config, PageFooter, PageType, Trail}
-import model.{ApplicationContext, Competition, CompetitionSummary, Group, Table, TeamUrl}
+import model.{ApplicationContext, Competition, CompetitionSummary, Group, StandalonePage, Table, TeamUrl}
 import navigation.{FooterLinks, Nav}
 import pa.{
   Fixture,
@@ -38,7 +38,6 @@ case class CompetitionMatches(competitionSummary: CompetitionSummary, matches: L
 case class MatchesByDateAndCompetition(date: LocalDate, competitionMatches: List[CompetitionMatches])
 
 trait DotcomRenderingFootballDataModel {
-  def filters: Map[String, Seq[CompetitionFilter]]
   def nav: Nav
   def editionId: String
   def guardianBaseURL: String
@@ -50,10 +49,10 @@ trait DotcomRenderingFootballDataModel {
 }
 
 private object DotcomRenderingFootballDataModel {
-  def getConfig(page: FootballPage)(implicit
+  def getConfig(page: StandalonePage)(implicit
       request: RequestHeader,
       context: ApplicationContext,
-  ) = {
+  ): JsObject = {
     val pageType: PageType = PageType(page, request, context)
 
     val switches: Map[String, Boolean] = conf.switches.Switches.all
@@ -302,6 +301,51 @@ object DotcomRenderingFootballTablesDataModel {
     Json.writes[DotcomRenderingFootballTablesDataModel]
 
   def toJson(model: DotcomRenderingFootballTablesDataModel): JsValue = {
+    val jsValue = Json.toJson(model)
+    withoutNull(jsValue)
+  }
+}
+
+case class DotcomRenderingFootballMatchSummaryDataModel(
+    footballMatch: MatchDataAnswer,
+    nav: Nav,
+    editionId: String,
+    guardianBaseURL: String,
+    config: JsObject,
+    pageFooter: PageFooter,
+    isAdFreeUser: Boolean,
+    contributionsServiceUrl: String,
+    canonicalUrl: String,
+) extends DotcomRenderingFootballDataModel
+
+object DotcomRenderingFootballMatchSummaryDataModel {
+  def apply(
+      page: MatchPage,
+      footballMatch: MatchDataAnswer,
+  )(implicit
+      request: RequestHeader,
+      context: ApplicationContext,
+  ): DotcomRenderingFootballMatchSummaryDataModel = {
+    val edition = Edition(request)
+    val nav = Nav(page, edition)
+    val combinedConfig: JsObject = DotcomRenderingFootballDataModel.getConfig(page)
+    DotcomRenderingFootballMatchSummaryDataModel(
+      footballMatch = footballMatch,
+      nav = nav,
+      editionId = edition.id,
+      guardianBaseURL = Configuration.site.host,
+      config = combinedConfig,
+      pageFooter = PageFooter(FooterLinks.getFooterByEdition(edition)),
+      isAdFreeUser = views.support.Commercial.isAdFree(request),
+      contributionsServiceUrl = Configuration.contributionsService.url,
+      canonicalUrl = CanonicalLink(request, page.metadata.webUrl),
+    )
+  }
+
+  implicit def dotcomRenderingFootballMatchSummaryDataModel: Writes[DotcomRenderingFootballMatchSummaryDataModel] =
+    Json.writes[DotcomRenderingFootballMatchSummaryDataModel]
+
+  def toJson(model: DotcomRenderingFootballMatchSummaryDataModel): JsValue = {
     val jsValue = Json.toJson(model)
     withoutNull(jsValue)
   }
